Remove BaseBindingBuilder.
