+++ formula-android/src/main/java/com/instacart/formula/android/internal/EmptyFragmentKey.kt
@@ -1,0 +1,9 @@
+package com.instacart.formula.android.internal
+
+import com.instacart.formula.fragment.FragmentKey
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+internal data class EmptyFragmentKey(
+    override val tag: String
+) : FragmentKey
