+++ formula-android/src/main/java/com/instacart/formula/android/internal/InternalExtensions.kt
@@ -1,4 +1,4 @@
-package com.instacart.formula.integration.internal
+package com.instacart.formula.android.internal
 
 /** Functionally the same as [Iterable.forEach] except it generates an index-based loop that doesn't use an [Iterator]. */
 internal inline fun <T> List<T>.forEachIndices(action: (T) -> Unit) {
