--- formula-android/src/main/java/com/instacart/formula/fragment/EmptyFragmentContract.kt
@@ -1,8 +1,0 @@
-package com.instacart.formula.fragment
-
-import kotlinx.android.parcel.Parcelize
-
-@Parcelize
-data class EmptyFragmentContract(
-    override val tag: String
-) : FragmentKey
