+++ formula-android/src/main/java/com/instacart/formula/integration/FragmentFlowRenderView.kt
@@ -9,6 +9,7 @@ import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import com.instacart.formula.Renderer
 import com.instacart.formula.RenderView
+import com.instacart.formula.android.internal.forEachIndices
 import com.instacart.formula.fragment.BaseFormulaFragment
 import com.instacart.formula.fragment.FormulaFragment
 import com.instacart.formula.fragment.FragmentEnvironment
@@ -16,7 +17,6 @@ import com.instacart.formula.fragment.FragmentFlowState
 import com.instacart.formula.fragment.FragmentLifecycle
 import com.instacart.formula.fragment.FragmentLifecycleEvent
 import com.instacart.formula.fragment.getFormulaFragmentId
-import com.instacart.formula.integration.internal.forEachIndices
 import java.util.LinkedList
 import java.util.UUID
