+++ formula-android/src/main/java/com/instacart/formula/integration/internal/CompositeBinding.kt
@@ -4,6 +4,7 @@ import com.instacart.formula.Evaluation
 import com.instacart.formula.Formula
 import com.instacart.formula.FormulaContext
 import com.instacart.formula.Stream
+import com.instacart.formula.android.internal.forEachIndices
 import com.instacart.formula.integration.Binding
 import com.instacart.formula.integration.ComponentFactory
 import com.instacart.formula.integration.DisposableScope
