+++ formula-android/src/main/java/com/instacart/formula/integration/internal/FeatureBinding.kt
@@ -5,6 +5,7 @@ import com.instacart.formula.Formula
 import com.instacart.formula.FormulaContext
 import com.instacart.formula.Stream
 import com.instacart.formula.android.FeatureFactory
+import com.instacart.formula.android.internal.forEachIndices
 import com.instacart.formula.fragment.FragmentKey
 import com.instacart.formula.integration.Binding
 import com.instacart.formula.integration.FeatureEvent
